Move the siteData function outside from routes and only call from the app bootstrap
